+++ src/main/java/ua/rd/pizzaservice/services/OrderService.java
@@ -4,5 +4,5 @@ import ua.rd.pizzaservice.domain.Customer;
 import ua.rd.pizzaservice.domain.order.Order;
 
 public interface OrderService {
-     Order placeNewOrder(Customer customer, Integer... pizzasID);
+     Order placeNewOrder(Customer customer, Long... pizzasID);
 }
