+++ src/main/java/ua/rd/pizzaservice/services/PizzaService.java
@@ -3,5 +3,5 @@ package ua.rd.pizzaservice.services;
 import ua.rd.pizzaservice.domain.pizza.Pizza;
 
 public interface PizzaService {
-    Pizza find(Integer id);
+    Pizza find(Long id);
 }
